+++ components/frontend/src/header_footer/Footer.js
@@ -7,11 +7,11 @@ export function Footer(props) {
             <Container>
                 <Grid divided inverted stackable textAlign='center' verticalAlign='middle'>
                     <Grid.Column width={5}>
-                        <Header inverted as='h4'>About <em>Quality-time</em></Header>
+                        <Header inverted as='h4'><em>Quality-time</em> v{process.env.REACT_APP_VERSION}</Header>
                         <List link inverted>
                             <List.Item as='a' href="https://www.ictu.nl/about-us">Developed by ICTU</List.Item>
                             <List.Item as='a' href='https://github.com/ICTU/quality-time/blob/master/LICENSE'>License</List.Item>
-                            <List.Item as='a' href='https://github.com/ICTU/quality-time/blob/master/CHANGELOG.md'>Changelog</List.Item>
+                            <List.Item as='a' href='https://github.com/ICTU/quality-time/blob/master/docs/CHANGELOG.md'>Changelog</List.Item>
                             <List.Item as='a' href="https://github.com/ICTU/quality-time">Source code</List.Item>
                         </List>
                     </Grid.Column>
@@ -20,9 +20,10 @@ export function Footer(props) {
                             <>
                                 <Header inverted as='h4' content="About this report" />
                                 <List inverted>
-                                    <List.Header>{props.report.title}</List.Header>
-                                    <List.Item>{props.report.subtitle}</List.Item>
-                                    <List.Item>{props.last_update.toLocaleString()} <span role="img" aria-label="little yellow diamond shaped separator">🔸</span> <em>Quality-time</em> v{process.env.REACT_APP_VERSION}</List.Item>
+                                    <List.Item>{props.report.title}</List.Item>
+                                    <List.Item>{props.report.subtitle || <span>&nbsp;</span>}</List.Item>
+                                    <List.Item>{props.last_update.toLocaleDateString()}</List.Item>
+                                    <List.Item>{props.last_update.toLocaleTimeString()}</List.Item>
                                 </List>
                             </>
                             :
@@ -37,7 +38,7 @@ export function Footer(props) {
                         <Header inverted as='h4' content='Support' />
                         <List link inverted>
                             <List.Item as='a' href="https://github.com/ICTU/quality-time/blob/master/README.md">Documentation</List.Item>
-                            <List.Item as='a' href="https://github.com/ICTU/quality-time/blob/master/USAGE.md">User manual</List.Item>
+                            <List.Item as='a' href="https://github.com/ICTU/quality-time/blob/master/docs/USAGE.md">User manual</List.Item>
                             <List.Item as='a' href="https://github.com/ICTU/quality-time/issues">Known issues</List.Item>
                             <List.Item as='a' href="https://github.com/ICTU/quality-time/issues/new">Report an issue</List.Item>
                         </List>